pyproject.toml
[project]
name = "rag_prompt_injection_detector"
version = "0.1.0"
description = "RAG-based prompt injection detector"
authors = [
    { name = "Mark Ikonnikov", email = "mrkikonnikov@gmail.com" },
    { name = "Strelkov Andrey", email = "astrelkov2004@gmail.com" },
    { name = "Prein Dmitry", email = "prein.ds@phystech.edu" },
    { name = "Korneenko Sergei", email = "Kornienko.sm@phystech.edu" },
    { name = "Portnov Timyr", email = "portnov.ta@phystech.edu" }
]
readme = "README.md"
repository = "https://github.com/asesen/rag-prompt-injection-detector"
keywords = ["data science", "machine learning", "rag", "prompt injection", "security"]
license = "Apache-2.0"

requires-python = ">=3.10,<3.12"
dependencies = [
    "langchain (>=1.0.7,<2.0.0)",
    "langchain-mistralai (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "faiss-cpu (>=1.7.4)",
    "numpy (>=1.26.4,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "torch (>=2.2.0,<2.3.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyarrow (>=15.0.0)",
    "fastparquet (>=2024.0.0)",
    "python-telegram-bot (>=22.5,<23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
